third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/access-control-and-redirects-async-same-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Tests that asynchronous XMLHttpRequests handle redirects according to the CORS standard.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src=/common/get-host-info.sub.js></script>
  </head>
  <body>
    <pre id="console"></pre>
    <script>
    var test = async_test("Tests that asynchronous XMLHttpRequests handle redirects according to the CORS standard.");

    function runTestAsync(url, credentials, addCustomHeader, expectSuccess) {
      xhr = new XMLHttpRequest();
      xhr.withCredentials = credentials;
      xhr.open("GET", get_host_info().HTTP_REMOTE_ORIGIN + url, true);
      if (addCustomHeader)
        xhr.setRequestHeader("x-webkit", "foo");

      xhr.onload = function() {
        test.step(function() {
          assert_true(expectSuccess);
          assert_equals(xhr.responseText, "PASS: Cross-domain access allowed.");
          nextTest();
        });
      }
      xhr.onerror = function() {
        test.step(function() {
          assert_false(expectSuccess);
          assert_equals(xhr.status, 0);
          nextTest();
        });
      }
      xhr.send(null);
    }

    var withoutCredentials = false;
    var withCredentials = true;
    var noCustomHeader = false;
    var addCustomHeader = true;
    var succeeds = true;
    var fails = false;

    var tests = [
      // Test simple same origin requests that receive cross origin redirects.

      // Request without credentials is redirected to a cross-origin response with Access-Control-Allow-Origin=*.
      // The redirect response passes the access check.
      ["/XMLHttpRequest/resources/access-control-basic-allow-star.py",
      withoutCredentials, noCustomHeader, succeeds],

      // Request with credentials is redirected to a cross-origin response with Access-Control-Allow-Origin=*.
      // The redirect response fails the access check because credentials were sent.
      ["/XMLHttpRequest/resources/access-control-basic-allow-star.py",
      withCredentials, noCustomHeader, fails],

      // Request without credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin.
      // The redirect response passes the access check.
      ["/XMLHttpRequest/resources/access-control-basic-allow.py",
      withoutCredentials, noCustomHeader, succeeds],

      // Request with credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin.
      // The redirect response passes the access check.
      ["/XMLHttpRequest/resources/access-control-basic-allow.py",
      withCredentials, noCustomHeader, succeeds],

      // Request without credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin
      // forbidding credentials. The redirect response passes the access check.
      ["/XMLHttpRequest/resources/access-control-basic-allow-no-credentials.py",
      withoutCredentials, noCustomHeader, succeeds],

      // Request with credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin
      // forbidding credentials. The redirect response fails the access check.
      ["/XMLHttpRequest/resources/access-control-basic-allow-no-credentials.py",
      withCredentials, noCustomHeader, fails],

    ]

    var currentTest = 0;

    function nextTest() {
      if (currentTest < tests.length)
        runTestAsync.apply(null, tests[currentTest++]);
      else
        test.done();
    }

    nextTest();
    </script>
  </body>
</html>
